EPMRPP-37572: Code review fixes - 1
