+++ app/src/pages/inside/dashboardItemPage/widgetWizardModal/widgetControls/controls/filtersControl/filterAdd/filterAdd.jsx
@@ -1,24 +1,54 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
+import classNames from 'classnames/bind';
+import { defineMessages, intlShape, injectIntl } from 'react-intl';
 import { connect } from 'react-redux';
-import { formValueSelector } from 'redux-form';
+import { reduxForm, formValueSelector } from 'redux-form';
 import { validate } from 'common/utils';
+import { FieldProvider } from 'components/fields/fieldProvider';
+import { FieldErrorHint } from 'components/fields/fieldErrorHint';
+import { Input } from 'components/inputs/input';
 import { messages } from '../common/messages';
 import { FILTER_ADD_FORM, FILTER_NAME_KEY } from '../common/constants';
 import { AddEditFilter } from '../common/addEditFilter';
-import { FilterAddFormBlock } from './filterAddFormBlock';
+import styles from './filterAdd.scss';
 
+const cx = classNames.bind(styles);
 const selector = formValueSelector(FILTER_ADD_FORM);
 
+const localMessages = defineMessages({
+  filterName: {
+    id: 'AddFilter.filterName',
+    defaultMessage: 'Filter Name',
+  },
+  placeholderFilterName: {
+    id: 'AddFilter.placeholderFilterName',
+    defaultMessage: 'Input filter name',
+  },
+});
+
+@reduxForm({
+  form: FILTER_ADD_FORM,
+  validate: ({ name }) => ({
+    name: (!name || !validate.filterName(name)) && 'filterNameError',
+  }),
+  onChange: ({ name }, dispatcher, { onChange }) => onChange({ name }),
+  enableReinitialize: true,
+  keepDirtyOnReinitialize: true,
+  updateUnregisteredFields: true,
+})
 @connect((state) => ({
   name: selector(state, FILTER_NAME_KEY),
 }))
+@injectIntl
 export class FilterAdd extends Component {
   static propTypes = {
+    intl: intlShape.isRequired,
     filter: PropTypes.object.isRequired,
     onSave: PropTypes.func.isRequired,
     onCancel: PropTypes.func.isRequired,
     onChange: PropTypes.func.isRequired,
+    valid: PropTypes.bool.isRequired,
     name: PropTypes.string,
   };
 
@@ -29,16 +59,34 @@ export class FilterAdd extends Component {
     onChange: () => {},
   };
 
-  handleFilterChange = (filter = {}) =>
+  getCustomBlock = () => {
+    const {
+      intl: { formatMessage },
+    } = this.props;
+
+    return (
+      <div className={cx('filter-add-custom-block')}>
+        <span className={cx('custom-block-text')}>{formatMessage(localMessages.filterName)}</span>
+        <FieldProvider name={FILTER_NAME_KEY}>
+          <FieldErrorHint>
+            <Input
+              placeholder={formatMessage(localMessages.placeholderFilterName)}
+              maxLength="128"
+            />
+          </FieldErrorHint>
+        </FieldProvider>
+      </div>
+    );
+  };
+
+  handleFilterChange = (filter) =>
     this.props.onChange({
       ...filter,
       name: this.props.name,
     });
 
   render() {
-    const { onCancel, filter, name, onSave } = this.props;
-
-    const customBlock = <FilterAddFormBlock onChange={this.handleFilterChange} />;
+    const { onCancel, filter, valid, onSave } = this.props;
 
     return (
       <AddEditFilter
@@ -46,9 +94,9 @@ export class FilterAdd extends Component {
         onCancel={onCancel}
         onSubmit={onSave}
         onChange={this.handleFilterChange}
-        canSubmit={Boolean(name && validate.filterName(name))}
+        canSubmit={valid}
         blockTitle={messages.addTitle}
-        customBlock={customBlock}
+        customBlock={this.getCustomBlock()}
       />
     );
   }
