+++ app/src/pages/inside/dashboardItemPage/widgetWizardModal/widgetControls/controls/filtersControl/filterAdd/filterAdd.scss
@@ -1,8 +1,8 @@
-.filter-add-form-block {
+.filter-add-custom-block {
   padding: 10px;
 }
 
-.form-block-text {
+.custom-block-text {
   display: block;
   margin: 0 0 5px;
   font-size: 13px;
