--- app/src/pages/inside/dashboardItemPage/widgetWizardModal/widgetControls/controls/filtersControl/filterAdd/filterAddFormBlock/index.js
@@ -1,1 +1,0 @@
-export { FilterAddFormBlock } from './filterAddFormBlock';
